refactor: 지하철 역 관련 API 예외처리 로직 추가 및 수정

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,40 +1,58 @@
 package wooteco.subway.station.application;
 
+import java.util.List;
+import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import wooteco.subway.exception.DuplicateStationNameException;
+import wooteco.subway.exception.NoSuchStationException;
+import wooteco.subway.exception.StationAlreadyRegisteredInLineException;
+import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 @Service
 public class StationService {
-    private StationDao stationDao;
 
-    public StationService(StationDao stationDao) {
+    private final StationDao stationDao;
+    private final SectionDao sectionDao;
+
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
+    @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
+        if (stationDao.findByName(stationRequest.getName()).isPresent()) {
+            throw new DuplicateStationNameException();
+        }
+
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
 
     public Station findStationById(Long id) {
-        return stationDao.findById(id);
+        return stationDao.findById(id)
+            .orElseThrow(NoSuchStationException::new);
     }
 
     public List<StationResponse> findAllStationResponses() {
         List<Station> stations = stationDao.findAll();
 
         return stations.stream()
-                .map(StationResponse::of)
-                .collect(Collectors.toList());
+            .map(StationResponse::of)
+            .collect(Collectors.toList());
     }
 
+    @Transactional
     public void deleteStationById(Long id) {
+        if (sectionDao.isExistStationId(id)) {
+            throw new StationAlreadyRegisteredInLineException();
+        }
+
         stationDao.deleteById(id);
     }
 }
